refactor: extract core code into a dedicated module
